+++ lib/celluloid/task.rb
@@ -1,6 +1,8 @@
 module Celluloid
   # Tasks are interruptable/resumable execution contexts used to run methods
   class Task
+    attr_reader :type # what type of task is this?
+
     # Obtain the current task
     def self.current
       task = Thread.current[:task]
@@ -14,7 +16,9 @@ module Celluloid
     end
 
     # Run the given block within a task
-    def initialize
+    def initialize(type)
+      @type = type
+
       actor   = Thread.current[:actor]
       mailbox = Thread.current[:mailbox]
 
@@ -40,5 +44,10 @@ module Celluloid
 
     # Is the current task still running?
     def running?; @fiber.alive?; end
+
+    # Nicer string inspect for tasks
+    def inspect
+      "<Celluloid::Task:0x#{object_id.to_s(16)} @type=#{@type.inspect}, @running=#{@fiber.alive?}>"
+    end
   end
 end