+++ spec/celluloid/io/actor_spec.rb
@@ -1,5 +1,0 @@
-require 'spec_helper'
-
-describe Celluloid::IO do
-  it_behaves_like "a Celluloid Actor", Celluloid::IO
-end