+++ spec/celluloid/io/mailbox_spec.rb
@@ -1,5 +1,0 @@
-require 'spec_helper'
-
-describe Celluloid::IO::Mailbox do
-  it_behaves_like "a Celluloid Mailbox"
-end